LPS-22219 Eliminate the needless usage of the Adaptor API
