--- portal-impl/src/META-INF/services/com.liferay.portal.kernel.adaptor.AdaptorFactory
@@ -1,1 +1,0 @@
-com.liferay.portal.adaptor.osgi.OSGiAdaptorFactoryImpl
