+++ portal-impl/src/portal.properties
@@ -5691,6 +5691,11 @@
     osgi.framework.storage=${liferay.home}/data/osgi
 
     #
+    # Set whether to register Liferay's services into the OSGi framework or not.
+    #
+    osgi.register.liferay.services=true
+
+    #
     # Set a comma delimited list of java packages that are exported to the OSGi
     # framework as bundle symbolic names.
     #
